+++ src/api/init.ts
@@ -40,7 +40,7 @@ import * as pathToRegex from 'path-to-regexp';
 import * as expressListEndpoints from 'express-list-endpoints';
 import { createMiddleware as createPrometheusMiddleware } from '@promster/express';
 import { createMicroblockRouter } from './routes/microblock';
-import { createTokenRouter } from './routes/tokenes/tokens';
+import { createTokenRouter } from './routes/tokens/tokens';
 
 export interface ApiServer {
   expressApp: ExpressWithAsync;
